requirements.txt
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.7.0
blurhash==1.1.4
boltons==23.0.0
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
exceptiongroup==1.1.2
face==20.1.1
flexicon==0.1.0
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
MastoBot.py @ git+https://github.com/e-dreyer/MastoBot/@717a8fbe5dd54de5fbc4fc307e599b62f54fa91c
Mastodon.py==1.8.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
prometheus-client==0.17.1
python-dateutil==2.8.2
python-magic==0.4.27
PyYAML==6.0.1
redis==4.6.0
regex==2023.6.3
requests==2.31.0
requests-cache==1.1.0
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
typing_extensions==4.7.1
url-normalize==1.4.3
urllib3==2.0.4
